requirements.txt
streamlit
boto3
strands-agents
nest-asyncio